# Projeto_Compilador/api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Eu Digital API
  version: 1.0.0
  description: API RESTful para gestão do “eu digital” — ingestão de pacotes SIP, CRUD de fotos, textos, resultados académicos, resultados desportivos, ficheiros e eventos.
servers:
  - url: /api
paths:
...
  /academicResults:
    get:
      summary: List academic results
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: institution
          schema:
            type: string
        - in: query
          name: author
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcademicResult'
    post:
      summary: Create an academic result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicResultInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicResult'
        '400':
          description: Validation error
  /academicResults/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get academic result by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicResult'
        '404':
          description: Not found
    put:
      summary: Update an academic result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicResultInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicResult'
        '400':
          description: Validation error
        '404':
          description: Not found
    delete:
      summary: Delete an academic result
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
...
  /files:
    get:
      summary: List files
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: author
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileItem'
    post:
      summary: Create a file record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileItemInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '400':
          description: Validation error
  /files/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a file record by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '404':
          description: Not found
    put:
      summary: Update a file record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileItemInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileItem'
        '400':
          description: Validation error
        '404':
          description: Not found
    delete:
      summary: Delete a file record
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
  /events:
    get:
      summary: List events
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: author
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation error
  /events/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get an event by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Not found
    put:
      summary: Update an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation error
        '404':
          description: Not found
    delete:
      summary: Delete an event
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
components:
  schemas:
    BaseItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        author:
          type: string
        type:
          type: string
        visibility:
          type: string
          enum: [public, friends, private]
        tags:
          type: array
          items:
            type: string
      required: [id, author, type]
...
    AcademicResult:
      allOf:
        - $ref: '#/components/schemas/BaseItem'
        - type: object
          properties:
            institution:
              type: string
            course:
              type: string
            grade:
              type: string
            scale:
              type: string
            evaluationDate:
              type: string
              format: date-time
    AcademicResultInput:
      allOf:
        - $ref: '#/components/schemas/AcademicResult'
      required: [id, author, type, institution, course, grade, scale, evaluationDate]
...
    FileItem:
      allOf:
        - $ref: '#/components/schemas/BaseItem'
        - type: object
          properties:
            originalName:
              type: string
            size:
              type: integer
            format:
              type: string
            description:
              type: string
    FileItemInput:
      allOf:
        - $ref: '#/components/schemas/FileItem'
      required: [id, author, type, originalName, size, format]
    Event:
      allOf:
        - $ref: '#/components/schemas/BaseItem'
        - type: object
          properties:
            title:
              type: string
            startDate:
              type: string
              format: date-time
            endDate:
              type: string
              format: date-time
            location:
              type: string
            participants:
              type: array
              items:
                type: string
            description:
              type: string
            eventType:
              type: string
    EventInput:
      allOf:
        - $ref: '#/components/schemas/Event'
      required: [id, author, type, title, startDate]